test(flow): consolidate i18n tests into single flow file
Merged three separate test files into tests/test_flow_multilingual.py:
- test_concept_views.py (10 tests, view CRUD + anchor + staleness)
- test_on_demand_attunement.py (3 tests, stub-backend enqueue flow)
- test_libretranslate_backend.py (6 tests, glossary post-substitution)

All 19 tests still pass; one file now carries the full multilingual
flow instead of three. Shared setup helpers live once at the top.
603 total tests, all green.

Spawning three small "one per implementation detail" test files for a
single user flow was the same dead-weight pattern the user has pushed
back on across the session. Saved as feedback so future additions
default to extending an existing flow file rather than creating new.
